[crater] Move target stuff into ::target module
I want to enforce some invariants on these types, which requires making
some of their fields non-visible, which requires them to be in a module.

Doing this in a separate PR for the benefit of clearer diffs for
subsequent PRs.
